Adding Market OrderType

+++ include/Usings.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <cstdint>
 
-using Price    = int32_t;
+using Price    = double;
 using Quantity = uint32_t;
 using OrderId  = uint64_t;
